requirements.txt
streamlit-mic-recorder==0.0.8
InstructorEmbedding==1.0.1
ctransformers==0.2.27
transformers==4.44.0
langchain==0.2.14
streamlit==1.37.1
pypdfium2==4.30.0
librosa==0.10.2.post1
pyyaml==6.0.2
langchain-community==0.2.12
chromadb==0.5.5